Move contentgrid mapping setup to its own configuration
