+++ contentgrid-spring-data-rest/src/main/java/com/contentgrid/spring/data/rest/webmvc/ContentGridSpringDataRestProfileConfiguration.java
@@ -1,16 +1,14 @@
 package com.contentgrid.spring.data.rest.webmvc;
 
+import com.contentgrid.spring.data.rest.mapping.ContentGridDomainTypeMappingConfiguration;
 import com.contentgrid.spring.data.rest.mapping.DomainTypeMapping;
 import com.contentgrid.spring.data.rest.mapping.FormMapping;
-import com.contentgrid.spring.data.rest.mapping.SearchMapping;
-import com.contentgrid.spring.data.rest.mapping.collectionfilter.CollectionFilterBasedContainer;
-import com.contentgrid.spring.data.rest.mapping.jackson.JacksonBasedContainer;
 import java.util.concurrent.atomic.AtomicReference;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.repository.support.Repositories;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.rest.core.config.RepositoryRestConfiguration;
 import org.springframework.hateoas.mediatype.hal.HalConfiguration;
 import org.springframework.hateoas.mediatype.hal.forms.HalFormsConfiguration;
@@ -19,6 +17,7 @@ import org.springframework.hateoas.server.EntityLinks;
 import org.springframework.hateoas.server.mvc.TypeConstrainedMappingJackson2HttpMessageConverter;
 
 @Configuration(proxyBeanMethods = false)
+@Import(ContentGridDomainTypeMappingConfiguration.class)
 public class ContentGridSpringDataRestProfileConfiguration {
     @Bean
     HalFormsProfileController halFormsProfileController(
@@ -29,29 +28,6 @@ public class ContentGridSpringDataRestProfileConfiguration {
     ) {
         var objectMapper = messageConverter.getObjectMapper().copy();
         return new HalFormsProfileController(repositoryRestConfiguration, entityLinks, domainTypeToHalFormsPayloadMetadataConverter, objectMapper);
-    }
-
-    @Bean
-    @FormMapping
-    DomainTypeMapping halFormsFormMappingDomainTypeMapping(Repositories repositories) {
-        return new DomainTypeMapping(repositories, JacksonBasedContainer::new);
-    }
-
-    @Bean
-    @SearchMapping
-    DomainTypeMapping halFormsSearchMappingDomainTypeMapping(Repositories repositories) {
-        return new DomainTypeMapping(repositories, (container) -> new CollectionFilterBasedContainer(container, 2));
-    }
-
-    @Bean
-    DomainTypeToHalFormsPayloadMetadataConverter DomainTypeToHalFormsPayloadMetadataConverter(
-            @FormMapping DomainTypeMapping formDomainTypeMapping,
-            @SearchMapping DomainTypeMapping searchDomainTypeMapping
-    ) {
-        return new DefaultDomainTypeToHalFormsPayloadMetadataConverter(
-                formDomainTypeMapping,
-                searchDomainTypeMapping
-        );
     }
 
     @Bean
